Fix crash introduced in PR #35
Pull request #35 was intended to simply silence "uninitialised value" warnings.
However, I, the author of that patch, misread which macro is used to compute the
value for `workspace_len` (in the PR description I allude to `Wosize_val`, but
link to the definition of a different macro.).  As a result, when `is_dfa` is
false, a NULL pointer read occurs.  This patch undoes this breakage;
`Wosize_val(v_workspace)` is now only evaluated on code paths where `is_dfa` is
true.